requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.8b0
cfgv==3.3.1
click==8.0.1
distlib==0.3.2
Django==3.2.7
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
identify==2.2.13
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathlib==1.0.1
pathspec==0.9.0
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
virtualenv==20.7.2
